added image preloader in css
Not sure if I like this, but I added an image preloader element in the
html file and loaded the image in the background in css. I think this
should speed up the drawing of the first block. It's probably better to
do this with javascript, but I'll change that when I get to it.