wrote some compatibility functions
added functions:
addClass: a method of Element.prototype that adds a class to the element
removeClass: a method of Element.prototype that removes a class from
the element
addListener: a method of Node.prototype that adds an event listener
removeListener: a method of Node.prototype that removes an event
listener

I will work on these functions in the future to make sure that they
behave properly in all modern browsers. Right now, they probably don't.